moved socket to game
